refactor: separate App from index
