+++ src/App.tsx
@@ -1,0 +1,34 @@
+import { CSSProperties, useState } from 'react';
+import clsx from 'clsx';
+
+import { Article } from './components/article/Article';
+import { ArticleParamsForm } from './components/article-params-form/ArticleParamsForm';
+import { defaultArticleState } from './constants/articleProps';
+
+import styles from './styles/index.module.scss';
+
+const App = () => {
+	const [articleState, setArticleState] = useState(defaultArticleState);
+
+	return (
+		<div
+			className={clsx(styles.main)}
+			style={
+				{
+					'--font-family': articleState.fontFamilyOption.value,
+					'--font-size': articleState.fontSizeOption.value,
+					'--font-color': articleState.fontColor.value,
+					'--container-width': articleState.contentWidth.value,
+					'--bg-color': articleState.backgroundColor.value,
+				} as CSSProperties
+			}>
+			<ArticleParamsForm
+				articleState={articleState}
+				setArticleState={setArticleState}
+			/>
+			<Article />
+		</div>
+	);
+};
+
+export default App;
